Refactored typeOfBusiness and otherTypeOfBusiness controllers to the common directory
